# country/src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Countries and Cities API
  version: 1.0.0
  description: Backend API for listing countries, finding cities by selected country, and retrieving city details.
servers:
  - url: http://localhost:8080
paths:
  /countries:
    get:
      summary: List countries
      operationId: listCountries
      tags:
        - Countries
      responses:
        '200':
          description: Countries returned successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Country'
  /countries/{countryId}/cities:
    get:
      summary: List cities for a country
      operationId: listCitiesByCountry
      tags:
        - Cities
      parameters:
        - name: countryId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
        - name: page
          in: query
          required: false
          description: Zero-based page index.
          schema:
            type: integer
            minimum: 0
            default: 0
          example: 0
        - name: size
          in: query
          required: false
          description: Number of cities per page.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          example: 10
      responses:
        '200':
          description: Paged cities returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CityPage'
        '400':
          description: Invalid pagination request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Country was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /cities/{cityId}:
    get:
      summary: Get city details
      operationId: getCityDetails
      tags:
        - Cities
      parameters:
        - name: cityId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: City details returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/City'
        '404':
          description: City was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
components:
  schemas:
    Country:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: India
    City:
      type: object
      required:
        - id
        - name
        - countryId
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Bengaluru
        countryId:
          type: integer
          format: int64
          example: 1
        population:
          type: integer
          example: 13193000
        zipCode:
          type: string
          example: '560001'
        description:
          type: string
          example: Technology hub and capital of Karnataka.
    CityPage:
      type: object
      required:
        - content
        - page
        - size
        - totalItems
        - totalPages
        - first
        - last
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/City'
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 10
        totalItems:
          type: integer
          format: int64
          example: 20
        totalPages:
          type: integer
          example: 2
        first:
          type: boolean
          example: true
        last:
          type: boolean
          example: true
    ApiError:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 404
        error:
          type: string
          example: Not Found
        message:
          type: string
          example: 'City not found: 999'
        path:
          type: string
          example: /cities/999